+++ api/src/types/index.ts
@@ -1,3 +1,3 @@
 export * from "./template";
 export * from "./graph";
-export * from "./chain-registry";
+export * from "./chainRegistry";
