+++ deploy-web/src/pages/validators/[address].tsx
@@ -4,7 +4,7 @@ import { useTheme } from "@mui/material/styles";
 import { makeStyles } from "tss-react/mui";
 import Layout from "@src/components/layout/Layout";
 import PageContainer from "@src/components/shared/PageContainer";
-import { BASE_API_MAINNET_URL, BASE_API_TESTNET_URL } from "@src/utils/constants";
+import { getNetworkBaseApiUrl } from "@src/utils/constants";
 import axios from "axios";
 import { ValidatorDetail } from "@src/types/validator";
 import { FormattedNumber } from "react-intl";
@@ -132,7 +132,7 @@ export async function getServerSideProps({ params, query }) {
 }
 
 async function fetchValidatorData(address: string, network: string) {
-  const apiUrl = network === "testnet" ? BASE_API_TESTNET_URL : BASE_API_MAINNET_URL;
+  const apiUrl = getNetworkBaseApiUrl(network);
   const response = await axios.get(`${apiUrl}/validators/${address}`);
   return response.data;
 }
